SONAR-2443 Remove indicator from suggest field to prevent "shaking"
FI: the display of the indicator is handle by the scriptaculous 
component, which uses the "display" property instead of the 
"visibility" one to hide/show the indicator. This is what causes 
the "shaking" of elements placed on the right side of the component
while waiting for the answer from the server.