SONAR-2073 - Fix which filters are visible and group them by category.
In a filter widget you shouldn't access to a filter created by another user but not shared by this user.

Moreover, in the dropdown list to select a filter to be displayed it would be valuable to display to different sections : one for "my own filters" and the other one for the "shared filters".