Specify data types on inserts into project_measures
At least when using Oracle databases, projects with large volumes of
updates (over a million inserts per hour) see lots of cursor
contention, accounting for up to 40% of the database's activity (after
adjusting the shared pool size appropriately). This is due to the
presence of null values where the database can't determine the data
type, and therefore has to try different variants of the statement.

Specifying the data types on the non-blob columns avoids the type
uncertainty, and solves the cursor contention.

Signed-off-by: Stephen Kitt <steve@sk2.org>