Execute Protocol Buffer compiler during build
Dependency for compiler selected based on platform. Supported platforms
are Windows/Linux x86/64 and Mac OS. Configuration in root POM to avoid
duplication.

This also makes version of compiler consistent with version of library.
Previously it was 3.0.0-a3 and 3.0.0-beta-1 respectively, now both are
3.0.0-beta-1. And removes all existing auto generated files and scripts
for their creation as well as their traces (such as comments and
exclusions for SonarQube).