Do not throw UnsupportedOperationException when httpRequest.getSession(false) is called
When httpRequest.getSession(false) is called, the callers does not ask to create a http session. (javadoc of HttpServletRequest: "If create is false and the request has no valid HttpSession, this method returns null.")

So throwing UnsupportedOperationException is not needed in this case and returning null is enough.

This will fix a [blocking issue](https://github.com/javamelody/sonar-javamelody/issues/4) in the [Sonar JavaMelody plugin](https://github.com/javamelody/javamelody/wiki/UserGuide#sonar-plugin).