SONAR-8458 store semantic version in sq-version.txt
Scanners load the file sq-version.txt to get the
version of SonarQube. Some scanners, for example Maven,
do not correctly support the new format on 4 fields.
For this reason a 3-digits format must be kept for
backward-compatibility. The complete version in the
new format is available in the new file
sonar-api-version.txt