[RFC] create issues for decreasing coverage
This is a RFC to start a discussion about a rule which triggers when the
coverage on a component decreases.

The implementation uses the magic parameter '0' to the existing coverage
rules to have a minimal impact on the existing code.
(And it does not make sense to set this parameter to '0')