Refactor Elasticsearch types "authorization"
Multiple indices define a type "authorization".
The related code was duplicated, and sometimes
had some minor differences. This commits share
all the management of these types in the same
bunch of classes.
It also allows to quickly create a project-related
index, without having to fix the different locations
which may require project re-indexing.