Force authentication by default
When new users initialise their SonarQube instance, it would be better to 
make their system to follow the 'secure-by-default' mantra. 

The current configuration of `sonar.forceAuthentication = false` means that 
as soon as a SonarQube instance is hosted and has scanned a project, there is 
potentially sensitive information exposed to non-authenticated users that could
just connect to the server.

While this is not an issue for more experienced users, for those less-knowledgeable, 
it can expose them to an unnecessary vulnerability.