SONAR-9909 Slow indexing of portfolios at server startup
The inner join is super slow because there's no index on column
copy_component_uuid. The join is a safeguard to ensure presence
of project reference. Dropping it is acceptable as long as the
model is supposed to be consistent. Any potential inconsistent states
will have to be fixed at the root.