Fix notifications of nightly builds
The branch "nightly-build" is replaced by the branch "dogfood-on-next"
every night by the Jenkins job "trigger-nightly-sonar-enterprise".
That triggers the execution of integration tests on all databases.

Unfortunately the team is not notified when tests fail. Errors are not
visible from:
- wallboards
- the project default page on burgr
- slack channels

The root cause is that burgr does not consider the branch "nightly-build"
as being shared by the team. The workaround is to rename the branch by
prefixing it with "branch-". This is convention for long-living
maintenance branches.