Add a CockroachDB dialect
This dialect allows users to run SonarQube with CockroachDB as a
backend. In general, CockroachDB is compatible with Postgres. Certain
settings and statements differ however. In particular, CockroachDB does
not currently support the `backslash_quote` setting, so this dialect
does not attempt to set it. Also, this dialect uses a JDBC URL prefixed
with `cockroachdb`.