Support space in //NOSONAR
Currently, SonarQube will suppress issues reported on lines ending with
a `//NOSONAR` comment, but does not affect lines ending with
`// NOSONAR`.

It looks like this makes it impossible to make use of this feature
together with code formatters that expect there to be a space in the
middle. In particular, Google's Java Style formatter always wants there
to be a space after the `//`, and deliberately provides no way to
configure this behavior.

https://github.com/google/google-java-format