SONAR-10485: Add minimum changeset size configuration for QG conditions
SONAR-9352 previous introduced conditions into the Quality Gate checks to ignore changesets that did not meet certain standards when those changesets did not change 20 or more lines. This bypass of conditions was hard-codes into Sonarqube Quality Gate checks so could not be disabled by an administrator, and effectively allowed technical debt to be introduced to a code-base through multiple small changesets that didnt contain appropriate coverage or duplicated existing lines.

This change alters the Quality Gate conditions to introduce a minimum number of lines the changeset much reach before that condition is viewed as effective, so that Sonarqube administrators can decide which conditions they may want to ignore, and what size the changeset is allowed to get up to before that condition starts being enforced.

To support this change, a new `minimum_effective_lines` database column has been added to the `quality_gate_conditions` table, with a migration setting this value to `20` for any metric that was previously checked in `SmallChangesetQualityGateSpecialCase`, and `0` for any other metric to ensure backwards compatibility. The default Quality Gate created on a new Sonarqube instance has also been altered to set the `NEW_COVERAGE` and `NEW_DUPLICATED_LINES_DENSITY` metrics to have a condition threshold of 20 lines for consistency with historical operations, however the other metrics that were historically skipped are not being set in the default quality gate since there were no historical default values for them.

The `create_condition` and `update_condition` web service endpoints have been altered to have an additional `minimumEffectiveLines` parameter on the request and response, with this being set as mandatory to ensure anyone using these endpoints doesn't accidentally break their instance configuration be overwriting customised configuration with default values.

Finally, the web interface has been altered to show the new condition limit fields in the Quality Gate configuration table, and allow editing of them through the conditions modal. The message that is shown on the scan overview page if any conditions have been skipped has also been altered to remove the reference to the hard-coded limits that previously existed, and make it clear that a configurable limit has been hit.