SONAR-11283: Add ability to exclude uncoverable lines in condition check
When committing changes that touched lines that could not have any tests performed against them - such as README files - these edited lines were included in the line count used to check whether the Quality Gate conditions should be applied, and therefore artificially caused the threshold on the condition to be met, so failed the Quality Gate if other code that could be tested but wasn't feasible to test was also included in the changeset.

This change alters the Quality Gate conditions to introduce a flag that controls what metric to use when calculating the number of changed lines that is then used when comparing against the threshold in each condition, thereby allowing Sonarqube administrators to decide which conditions they may want to ignore if multiple lines are changes in non-code files, but only a limited number of executable lines are changed, or whether all edited lines should count towards the threshold. For backwards compatibility this new flag defaults to false so treats all edited lines as counting towards the threshold.

To support this change, a new `only_coverable_lines` database column has been added to the `quality_gate_conditions` table, with a migration setting this value to `false` for all existing conditions to ensure backwards compatibility. The default Quality Gate created on a new Sonarqube instance also sets this field to `false` for consistency with historical setups.

The `create_condition` and `update_condition` web service endpoints have been altered to have an additional `onlyIncludeCoverableLines` parameter on the request and response, with this being set as mandatory to ensure anyone using these endpoints doesn't accidentally break their instance configuration be overwriting customised configuration with default values.

Finally, the web interface has been altered to show the new condition  fields in the Quality Gate configuration table, and allow editing of them through the conditions modal.