SONAR-14792 Fix issue on Oracle DB by recreating PK for 'ce_activity'
When switching from ID to UUID for 'ce_activity' table, it has been forgotten to drop unique index on UUID, though PK has been created. In all DBs it worked well, but for Oracle already existing index has been associated with PK (worked well on upgrade path). When new schema has been introduced, that's not the case, as for Oracle when creating PK, unique index also will be created. That means we tried to create another index on top of PK index, which was causing failures.

To fix that issue previous UUID index will be dropped and PK will be recreated, so that we end with clean state.