Change log level of merge base sha1 to info
This info is too important to be hidden behind debug log level. I've just spent days trying to figure out why SQ doesn't detect new code properly, and it turned out to be because of some misconfiguration which I wouldn't have been able to figure out if not for this piece of info.