SONAR-21575 Populate the issue filter chain after optional plugins have been loaded
Optional plugins can contribute IssueFilter, so we should not load the list of extensions too early.
Moving the IssueFilters class in the child container was not possible due to many dependencies, so the choice has been made to make it "mutable": it is instantiated early to allow injection, but the list of IssueFilter is populated later.