SONAR-24376 Rewrite ProcessWrapper using Apache commons-exec
Maintaining a custom implementation started to be quite complex, and error prone. The initial use of this class
was to run simple commands, getting stdout synchronously.
But when used for more complex commands, it started to show its limits:
  * dead lock when not consumming stdout/stderr buffers
  * poor management of exceptions

Using commons-exec will also allow us to set timeout if needed.